pyproject.toml
[tool.poetry]
name = "shop-bot-server"
version = "0.1.0"
description = ""
authors = ["Eldos <eldos.baktybekov@gmail.com>"]
readme = "README.md"
packages = [{include = "shop_bot_server"}]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.4"
django-environ = "^0.9.0"
djangorestframework = "^3.14.0"
Pillow = "^9.3.0"
django-admin-rangefilter = "^0.9.0"
psycopg2-binary = "^2.9.5"
coinbase-commerce = "^1.0.1"
celery = "^5.2.7"
django-celery-beat = "^2.4.0"
redis = "^4.4.0"
httpx = "^0.23.3"


[tool.poetry.group.dev.dependencies]
coverage = "^7.0.3"
django-silk = "^5.0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
